Delete existing unwanted network interfaces (#11)
Delete existing unwanted network interfaces but only if it was
configured via systemd-networkd, i.e., only if the netdev config
file is about to also be deleted.

Note that we delete the interface first and the netdev config file
second. This is so that, in case deletion of the interface fails,
a rerun of the role attempts to delete it again.